Fix ST hyperlink rerouting on click
Hyperlinks with a non-URL URI would result in the web-app rerouting
instead of opening the hyperlink dialog for editing. This is probably
because of Angnular ui router. With this change we only add URLs to the
hyperlinks `href` in slate (so that the user can right-click > open in
new tab).