[AUTH-412] Update heading RegExp to conform to CM
Links `marked-ast` versioning to our local fork, which updates the Markdown
editor's RegExp for headings (as used in the preview functionality).
Requires a space between the hash and the heading markup, which conforms to
the CommonMark spec and is necessary for GitHub compatibility.

See:

- https://github.com/github/markup/issues/1013#issuecomment-287060641
- https://blog.github.com/2017-03-14-a-formal-spec-for-github-flavored-markd

Addresses this support ticket:

https://contentful.zendesk.com/agent/tickets/29681

Addresses this JIRA ticket:

https://contentful.atlassian.net/browse/AUTH-412