Push assets from all builds to prod S3 bucket
Currently, due to user_interface/pull/3312, you can deploy any commit
to all our environments EXCEPT production. This limitation is because
the assets are pushed to our production S3 bucket only when a commit
is built on master.

This removes that restriction, thus enabling us to truly deploy any
commit to production.