Fix space usage page
A child component was changed to use the service provider, which worked fine in org usage page, but not here in space usage page.
This is because the service provider is added to the react context by the react-component directive, which was not used for the space usage page.
This change migrate the page to use the react-component directive, something we want to do anyways.