Applies validation on rich text hyperlink modal (#3531)
* Applies validation on rich text hyperlink modal

This applies rich text field validations on the link types of the
hyperlink modal. It removes the URI link type as a default to handle
cases when it is not included in the validation.

* Renames allowedHyperlinkTypes -> entitySelectorConfigs

* Adds allowedHyperlinkTypes prop

This adds a new prop to HyperlinkDialog component that sets the allowed
hyperlink types based on links validation settings. This also shows the
correct link type when user is editing a link which type is no longer
valid.

* Mocks LINK_TYPES and uses Visible component for link type select options:

* Moves Visible component and bumps validation version