Track rich text preview for all locales at once
Instead of adding one "feature_text_editor" context event for
each rich text field locale to the "element:click" event with
"openContentPreviewBtn" element ID, we now only add one context
event per RichText type field.
This is necessary as the total payload to Snowplow can easily
become so big that it will be rejected, thus, we would end up
without the knowledge that a user clicked on the preview button
on a rich text entry.