Sync token data directly to redux store
Replaces listener on scope and therefore dependencies on angular.
Looses the environments, as that doesn't seem to be part of the token.
No need to fix downstream selectors, as this part of the state is not used yet.

Also fix warning of my IDE, which I tend to agree with (local variable is redundant).