Sends compact OT ops for RichText (#3679)
* Sends compact OT ops for RichText

OT operations that change node types result in an error in Sharejs
where the nodeType property contains the string "PERMISSION_CHECK_STRING" which
renders it invalid. This happens when doing string deletion (changing
the nodeType property), so to avoid this we send a compact operation
that does object deletion/insertion instead.