Show error locales even if localization is disabled
If we do not show fields in locales that are existing on the entry (e.g.
because localization was previously enabled) in case of an error, there
is no way for the author to resolve the validation issue, plus there is
not even an error message. This is a pragmatic fix to solve the issue
for the author by showing the failing locale and its error message.