Fix LD not initializing on slow connections
This was occuring due to a race condition. On a slow network,
the `contentPreviewsBus$` stream emitted an error before LD was
successfully initialized. This caused `userDataBus$` to fail and not
call its handler which actually is incharge of switching context in LD
to the current user and initializing LD.

Basically, if LD wasn't initialized in the first 2500ms of app
load (since that's the poll interval for `contentPreviewBus$`), it
would never initialize due to the aforementioned reason.