[AUTH-819] Fix "Embedded Entry Inline" button
Fixes an issue that prevents the "Embedded Entry Inline" button from being
used. From the JIRA ticket:

> When the ‘Embed Inline Entry’ rich text editor option goes into button,
  rather than dropdown, form (which happens when both embedded block entries
  and embedded assets are disabled in the content model), clicking the
  button does nothing. In the editor, it raises a ‘missing handleClick’
  error related to the component implementation.

The reason is because at some point we switched to using a mouse-down event,
rather than a click event, for the embedded entry inline component, but
neglected to update the handler property for the button (so now only the
dropdown option works).

This also swaps another instance of handleMouseDown in the embedded entity
block toolbar icon (which handles both embedded asset inlines and embedded
entry blocks) for handleClick. Following discussion with Sergii, we don't
need to use handleMouseDown for any of these handlers/callbacks anymore -
this was a hack to bypass issues with browser focus shifting from the rich
text editor to the button. This should no longer be an issue, so we should
prefer the more conventional (and better supported) handleClick in every
case.