[AUTH-749] Don't display disallowed asset option
Disables the "Embedded Asset Block" toolbar option (in dropdown and button
form) when the user lacks sufficient permissions to access assets.

Does so by propagating the AccessChecker 'section visibility' value for
assets as permission properties through the widget API. If the asset section
is not visible to the user (i.e., if the user policy disallows asset
access), they should not be able to view the asset embed option in the
rich text editor toolbar, either.