Fix slug widget auto-revert regression
When editing slug fields that track the current title, the change (e.g.
deleting the slug field value) would immediately be reverted. This is
a regression not caught by our tests, possibly because we don't have
integration tests for this and the widget API changes. More
investigation is needed to find out what exactly caused this bug.