[AUTH-852] Fix ShareJS error on rich text clear
Fixes an issue in which ShareJS outputs a type error to the console when the
rich text field state is cleared (does not apply to newly created
documents). This arised as a consequence of the implementation of AUTH-602.
The underlying issue is that we are passing the current field state to be
diffed against undefined, which ShareJS doesn't know how to handle (it's
expecting to diff an object against another object). Hence we should send
both values as undefined. This is a purely technical change which does not
affect UX.