Fix RT version bump on document with embeds/links/hr
Consider insertion of empty text node as an "irrelevant" operation by
Slalte. That means that on such an operation we won't persist the change
and thus prevent a version bump with effectively no change to the
document (the empty text node would be lost in the conversion to the CF
document anyhow).