Use slide-in navigation for bulk editor
This change refactors the bulk editor to make use of our slide-in
mechanism. Bulk editor can now be slided-in not only from the root slide
but also from a slide on a higher level of the slide stack. It also
means that the bulk editor now behaves like entry/asset slides, meaning
that e.g. the peek feature is working with it.
Clicking another link from within a bulk editor entry will now
slide that entry or asset in on top of the bulk editor rather than
opening it in a separate editor view with a fresh slide stack.

On a more technical level:
 - Registers `BulkEditor` as a new slide type to
   `navigation/SlideInNavigator`. For the `previousEntries`
   parameter we use `<entryId>:<fieldId>:<locale>:<referenceIndex`
   to identify a bulk editor slide.
 - `EntityPageController` is now responsible for loading all slided-in
   editors `editorData`.
   - This means that this logic got removed from the
     `cfEntryEditorDirective` and the `cfAssetEditorDirective`
   - It also allows us to more efficiently load and share the bulk
     editor's as well as the parent entry's `editorData` in one go.
   - This will easily allow us to defer loading of background slides
     on a page load in the future.
 - Removes `widgetApi._internal` and frees the reference editor widget
   from the responsiblity of constructing any of the bulk editor's
   dependencies.

This also:
 - Fixes `feature_bulk_editor` event tracking (see newly created
   docs/analytic_events/feature_bulk_editor.md for detail).