Get rid of contextHistory in Entry/AssetController
Also solves a slide-in focus bug caused by a race condition where
a slide not on top that would load later than the top slide would
steal the top slide's focus on an initial page load.
This bug still persists in case of the bulk editor which is still
stealing the focus from any slide above.