Remove breadcrumb "..." from entry versions view
The "..." is buggy as we do not use the legacy breadcrumbs anymore for
the entry editors, instead we use slide-in.
This has been buggy for a while so it makes sense to just entirely
remove this. This was also the last occurence of this feature so it's
gone entirel from our code base. We are still using the legacy
breadcrumb "<" in some places (mostly settings views).