Fix magic version bumps on RT document with marks
As our adapter produces a valid SlateJS document but not a fully
normalized one, the editor is doing some normalization on that value
when instantiated. As the operations triggered during this normalization
are not well distinguishable from operations that would be triggered by
user actions, this would cause us to trigger a write operation reulting
in a magic version bump.
Instead of leaving the normalization to the Editor instance, we now
normalize the value before instantiating Editor.
While ideally the adapter would return a fully normalized document, this
would be more effort also in the long term as we had to update it every
time Slate changes a detail on its document format.