Make new localeData approach work with bulk editor
Also refactors the setLocaleData util to export a
named function that just hydrates $scope.localeData,
initializing it as well if necessary. This is nice
for the bulk editor since we have no entry sidebar
niceties to worry about (and we always want
single-locale mode to be off). Also cleaned up
the file as a whole.