Entry activity sidebar widget (#4072)
We add a built-in sidebar widget that saves and renders the entry activity from Stream. Entry content changes are logged only if the user leaves a field after modifying it.

We add a new event bus in Document.es6 that gets updated when the following events happen:
* Field is unfocused
* Entry status changes
* Entry content changes

We add a new sidebar event type called UPDATED_DOCUMENT_STATE that is emitted only if the field is now blurred and was modified or if the event is neither blur nor change, which means status change.
We add a sidebar widget component that subscribes to SidebarEventTypes.UPDATED_DOCUMENT_STATE and saves the events to Stream. It also subscribes to changes in Stream and renders the feed as entry activities.