Provide noop loadEvents to widgets
Currently when loading an entry with rich text field in the bulk editor,
the field is not rendered and there's an error in the console.
We fix this by always providing a `loadEvents.emit` to widgets, either
the real one or a noop one like in the context of the bulk editor.