Fix link field empty value resulting in array
An empty link field should result in `undefined` rather than an empty
array.
This was actually the case in the CMA after removing the last link, but
somehow a min/max number of links validation was failing after the last
link of a link field was removed, possibly because in the ShareJS
document the value is `[]` rather than `undefined`.