Remove timeout for wait-on
As of #4484, we can be sure that the server will be reported as ready
only after all assets are compiled. Now that still takes random amount
of times on CI based on the load on CI. Adding this delay param makes
contract tests flaky as for some runs the server comes up in under 60s
and in some runs it doesn't. Removing this ensures that wait-on
resolves whenever the server is ready.

Also, this might potentially cause a build to hang if the server
doesn't come up which I believe is a good thing as that's should point
us to a problem we need to debug.