Persist all task specific comment properties
We no longer tread comments as tasks in the tasks feature. Meaning we
remove all code mocking additional properties on top of comments. We
also persist all task (or "assignable comment") specific information
even though the backend does not yet support it.
This may temporarily result in some weird UI (especially considering the
backend does not throw errors in all expected cases) for this WIP
feature which is not yet enabled for any customers.