Add memoization
Note that all dependencies have to preserve identity of their return value for memoization to work.
That's why dependent selectors that didn't preserve identity have been memoized as well.

Also note that the cache will get invalidated if `getOrgRole` is called with a different orgId, even if it's called from another selector.
This could be prevented by changing `getOrgRole` to return a map of roles for each org instead of using a prop attribute.