Revert filestack update from 3.6 back to 1.14
Some users have experienced issues when uploading certain file types
(.obj, .manifest, etc.). We were able to reproduce the issue on some
machines but it works on others. Will have to investigate this further.

The issue is related to the ShareJS ot.contentful.com call not
containing a `file[locale].contentType`. The connection to Filestack is
not clear here as the filestack upload does contain a `mimetype`.