Migrate all Jade templates to HTML
- Create HTML loader for Karma

This way we can have the same functionality in both Karma and Webpack.

- Remove beforeEach/afterEach from DSL

Adding a new loader to SystemJS causes the execution/import code to run a bit
slower than before, but beforeEach in the DSL essentially imported and executed
$apply every 10ms, meaning that if Angular was instantiated after the first 10ms
interval, angular-mocks would throw an error stating that the injector has already
been created and cannot register the module. Instead of using a hammer to force
all promises to work via $apply, a new function `$waitFor` is used which duplicates
the interval functionality but on a per-promise level.