Add createReadOnlyWidgetAPI() for read-only RT
Since in the versioning view we are relying on an actual RichTextEditor
instance, it makes sense to inject an actual widgetApi rather than a
monkey patched object like we used to with the old version of the editor
while hoping that it won't break when the editor changes.

This change also does some mild refactoring on `createNewWidgetApi()` to
make dependencies more transparent rather than just injecting `$scope`.