[AHOY-66]: add our own styling for the cookie banner and management panel (#5594)
* feat(cookie-consent): remove unecessary elements added by Osano

Osano add some buttons and links that we won't use to the cookie consent banner. And also does not allow us to change the banner text. So we did it programatically

* feat(cookie-consent): reset some css styles

We need to overwrite some css in the cookie consent banner to apply our own styles instead of the vendor's

* feat(cookie-consent): improve button styles and add manage consent button

* chore(cookie-consent): update css and revert toogle

* refactor(cookie-consent): improve banner styles and remove Marketing toggle

* refactor(cookie-consent): show checkbox instead of toggles

* refactor(cookie-consent): changed button to be primary

* chore(cookie-consent): remove osano setup

* feat(cookie-consent): improve button styles and add manage consent button

* feat(cookie-mgmt): new styles for the cookie management slide (#5614)

This overwrites Osano's styles in the cookie consent management slide with our own styles

* chore(cookie-consent): move cookie styles to a separated file

* refactor(cookie-consent): check if toggle exists before hiding it

Osano takes some seconds to load the banner. If we try to hide the banner before it shows up in the screen, we get an error

* feat(cookie-consent): hide  cookie widget in the bottom right of the screen

* refactor(prelude.js): revert changes from cookie consent styling story

* fix(cookie-consent): hide marketing toggle in management panel

* refactor(cookie-consent): change management panel styles