fix(editor): do not track CmaDocument edit conflict on bad requests  (#5778)
Previously, because of the potential handling of deleted entity kind of edit conflicts,
we might would have tracked some bad request cases as edit conflicts where the
`BadRequest` was not due to entity deletion. E.g. in case of an archived entity.
We are now stricter on the tracking to avoid these false positives that are not
actually edit conflicts.