fix(editor): disable fields on archived entity in CmaDocument editor (#5810)
We were disabling fields in case of an archived entry whenever the
OtDocument was disconnected. It is also disconnected in case of archived
entry as in this case a ShareJS connection isn't possible. In case of
CmaDocument , we are always "connected" by design (unless internet is
down) so we need to be more granular on the state.