test(cmadocument): fix incomplete test on CMA taking more than 5s (#5903)
This test was changed in 6591f4f for some reason while it made perfect sense
the way it was. The final assertions verifying the second API call are crucial for
this test as otherwise we do not assert at all what the test is all about - ensuring
that changes during a long save operation will trigger a subsequent save. This
assertion and the whole test are very important as they have to fail if anyone
were to "optimize" the save interval by e.g. removing the `.merge(afterSave$)`

This also fixes the `collects changes made during saving and sends them only
5s after last CMA request` test as it also didn't make much sense (using a
separate `setValue` call to trigger a save rather than relying on the same logic
that would ensure that we persist any changes made during a save subsequently.