fix: Map fields in SpaceTemplateCreator
Fixes a bug that manifested in the EnterpriseSpaceWizard refactoring. The template list
component was given templates that had fields that were not  mapped correctly, and so it did not show
the templates properly. This moves all field mapping logic into the service and updates all code that
uses the service.