chore: Use gitSha instead of version
At some point between versions 4.8.1 and 15.3.1, `version` became an option
that you cannot override in `yargs`. In this case, if you try to override
`version` it silently ignores it and continues to keep `version` as a boolean.
This led to our build failing because `version` became false, and false is not
a string, causing the path generation to fail when creating staging/preview/prod
distributions. Additionally, this was not clear in their documentation or the
generated help; the docs don't reference this behavior and the generated help
looks ok:

Options:
  --version, -v  Set the version identifier                  [string] [required]
  --branch, -b   Branch name                                 [string] [required]

When in reality, version is not overwritten and is a boolean. The output now looks like this:

Options:
  --version         Show version number                                [boolean]
  --branchName, -b  Branch name                              [string] [required]
  --gitSha, -g      Git commit SHA hash                      [string] [required]

This changes version to gitSha, and branch to branchName to be slightly more explicit
and to work around this.