fix(modal-launcher): unmount (#6216)
* fix(modal-launcher): unmount

The current implementation is never unmounting the given component.
Once mounted react takes care of the container element. calling `remove` has no effect as long as the react component is mounted. React will take care the container of a mounted component exists.

With the changes of this PR we now explicitly unmount the given component and remove the container afterwards.
The `delay` is needed to show our `Modal`s hide animation.