fix(CmaDocument) Track resolved asset file conflicts
As of 7b169dc we are merging remote entity changes back into the local state of `CmaDocument`
so we should also count remote changes to an `asset.file` as a potentially succssfully resolved
conflict if applicable. This case was not relevant previously as we were only expecting the
asset processed event where a file change was expected and we had the assumption that we did
not want to track this case.

This change also iterates on this assumption and updates the code
to also track successfully resolved changes due to a processed file caused by the asset system.
Motivation for this are:
A) Simplified logic and more generic handling and conflict tracking in all cases.
B) The fact that we were also tracking unresolvable conflicts in this case already.
C) The necessity that we would have to add even more complex logic to further check whether
   there are any remote changes to asset files not due to the asset system.
D) Haveing more complete data on prevented conflicts no matter cause (basically considering the
   asset system a background actor making changes indirectly triggered by the user)