fix: Use local version of wait-on
See https://github.com/lodash/lodash/issues/4837 and https://github.com/lodash/lodash/issues/4843

Lodash broke in version 4.17.16 (specifically /fp is missing) and wait-on depends on lodash/fp. To
mitigate this from potentially happening again (this seemed to happen due to a shared lack of
understanding of the lodash build/release process...) wait-on is moved to the repo and as an npm
script, rather than globally.