improvement: Add fallback values when flag variation retrieval fails
Handles any errors, expected or not, that come from the flag variation
retrieval call, and return the fallback value in those cases.

- Refactor to cache based on flag in addition to org, space, env ID
- Change `client.allFlags` to `client.variation`

This will help give better insight into which flags are active in our app
as well, as the current implementation simply marks all flags as active
(even if we don't use them) in the LD dashboard.