feat: don't rewrite secure asset urls
In order to implement secure assets, we need the backend to rewrite
the secure asset urls when coming from the CMA (it does this ignoring
the skip transformation flag only for the url value of the assets).
The frontend should allow these values through as-is.

These urls look like:

  (images|...etc).secure.(ctfassets.net|flinkly.com|quirely.com)

This shouldn't break any existing flows that we're aware of since all
of the other urls returned from the CMA shouldn't have `.secure.` in
them.