feat: Handle app mgmt validation errors (#6760)
* feat: add app mgmt validation messages

* feat: placeholders and no fallback in nav preview

* feat: show validation errors for page location

* fix: use min-content as page location height

With validation messages the content of page location did overflow. With
min-content the div height adjusts automatically depending on the
content.

* feat: add frontend validation for app mgmt

* feat: use proper generic error message

* feat: show backend errors in app ev frontend

* feat: remove validation error on input change

* test: add tests for AppDetails validation

* fix: allow localhost app src

* empty commit to trigger ci

* fix: don't use punctuation in messages

* fix: red location validation messages

* fix: location config transition

It is only possible to transition between fixed values. Not between
dynamic ones like (min-content). Therefore we now don't transition the
height but the maxHeight as we roughly know what the maximum height of
the location config can be.

In addition to that the transition of padding-left/right was removed as
it looks strange.

* feat: validate fieldTypes in entry-field location

* style: make code more consistent

* style: run prettier