perf(content-tags): flag based rendering (#6872)
* perf(content-tags): flag based rendering

With the old approach, we had (in several places) a conditional rendering based on the value of the tags feature flag. With the feature flag being async, a changed flag value would always result in a different tree of the condition and an unmounting of the old component tree and mounting of the new component tree based on the condition. This resulted in unnecessary requests.

We now apply the condition directly in the tags provider, and only return an empty array for tags if the feature is disabled.

* fix failing test