feat(chore): increase node version (#6442)
* feat(chore): increase node version

set docker node version to same version as we use in nvmrc for local testing.

* fix(tests): retry timers

- remove obsolet dependency `jest-dom` since we use it's successor `@tsting-library/jest-dom`
- make sure all timers are reset correctly
- make sure we use a new instance of `withRetry` for every test since it has it's own `setTimeout` state (not sure why it worked before)

* fix(tests): snapshots

For some reason, with node 12 the result for `sorts by status` has been off by 2 (`15` instead of expected `17`).
Looking at the resulting list, we get `19`, `17`, `15`, `13` and so on. I changed the test to expect the row at index `1` (instead of `2`)
to be of index `17`.

In addition, I also update deprecated API usage.