fix: broken bulk editor "Create new entry and link"
Fixes the bulk editor action to "Create new entry and link" which seems to
be broken with some recent refactoring. It seems the issue is related to us
using the `APIClient` for fetching the entry while we used to use the legacy
client returning the entry as returned by the CMA with
`x-contentful-skip-transformation` header.

Note: It's uncertain where this got broken, #6777 or #7022 don't seem to be
the cause for this. The tracking event `entry:create` for `eventOrigin:
"bulk_editor"` actually got broken around July 2020 and should be fixed
with this change.