fix: cmaDocument#removeValueAt() producing empty values in arrays
Using _.unset(), cmaDocument#removeValueAt() did leave some empty values
in arrays when removing a value. This caused a bug in the bulk editor
where removing a reference from within bulk editor would result in an
invalid entry, causing the CMA to respond with an error when CmaDocument
persists the change.
In the OtDocument version using ShareJS this has never been an issue
despite the empty value as it seems this has been handled by ShareJS on
the backend.