ci: Disable Saucelabs runs
The e2e tests, and therefore the Saucelabs runs, seem to be broken. Since we
disabled the other e2e tests, and there's work planned to add monitoring for
the web app, this should be ok for now, and it can be revisited at a later
point if we want to re-enable the Saucelabs runs.