feat(analytics): track v3 of feature_reference_action event to Segment (#6649)
Updates data tracked for the `feature_reference_action` which was previously defined as v2-0-0 in
Snowplow and which is now defined in Segment as v3. This adds additional information to reference
editor reference interactions, mainly referenced entity IDs and it fixes some events in an media
editor (asset links) context.