feat(content-tags): ensure single provider in hierarchy (#7576)
As a performance optimisation, we want to ensure a single `TagsRepo` (not so relevant), and a single `ReadTagsProvider` (very relevant) for one react root.
ReadTagsRepo is the single source of truth for tags in our frontend code, and should only be available once. It also caches all loaded tags for the session.