fix: field `unique` vallidation checkbox state (#7612)
When setting a CT field's uniqueness validation via the CMA like

validations: [{ unique: false }, ...]

it would result in the field editor UI showing it as enabled while it's
actually set to `false`. That's because the UI would create and assume
an empty array `[]` in this case rather than explicitly checking whether
this validation is set to `false`.
The `unique` validation seems to be the only affected validation as all
other validations come with more complex validation settings, e.g.

{ size: {
  min: 42, max: 420
} }

so there's no other way than removing them from the array altogether
to deactivate them.