Remove 'onboarding deployment completed' Intercom tracking
This event must have been broken since d8f5417 as we're receiving
`data.data` and a transformed `element_id` instead of `elementId` now.

Apparently the event got archived in Intercom already, so we should
be able to safely delete it here instead of re-implementing.
If we ever want to get this back we should implement it via
`Intercom.trackEvent()` instead of relying on this kind of hack
anyways.