refactor: correct release Segment event registration
Fixes how Segment only "release:" web app events are registered. It
seems there never was a Snowplow schema in production, so no reason
to register and track the event for Snowplow in the web app.

Unfortunately, `release:dialog_box_open` and `release:dialog_box_close`
web app events were tracked against these same Segment schemas that do
not actually exist. The correct schema for both these events would have
been `release_dialog_box` which is used for both events. This PR does
not change this as it would result in data going into different tables
(named after the schema) which should be done in a follow-up in
coordination with the team.