feat: `environment_key` for most Segment events
Updates `Analytics.trackContextChange()` to allow setting the
environment. This allows us to attach the `environment_key` to
all events in the Segment tracking that also have a `space_key`.

We can't do the same for Snowplow events as this would require a
schema change for all events, which wouldn't be worth the effort
as we are in the process of retiring Snowplow.