refactor: remove `app_open` event contexts
Removes the `contexts` from events tracked for the `app_open`
Snowplow/Segment schemas and instead adds the default event props like
space ID to the main event.

This will break the Snowplow event as it doesn't account for these
props. We could either update the schema to `1-0-1` or stop tracking the
event since we're switching to Segment anyways.