refactor: remove `entry` event context
Removes `entry` Snowplow contexts event from `entry:` events and in turn
introduces a `entry_version` and `content_type_id` field on the main
event.
This leaves the `entry:create` context without any context events and
the `entry:publish` event 1 `sidebar_render` and 0…* `extension_render`
context events.

This will break the Snowplow events as it doesn't account for the two
new props. We could either update the schema to `2-0-2` or stop tracking
the event since we're switching to Segment anyways.