fix: [] Convert document to internal locale code (#8389)
There was an issue with new new implementation of the entry editor sdk that the fields value are not showing up.

In the old implementation the path was always generated with the internal_code of the locale. The new implementation was using the normal "public code". That's why with the new sdk the values were grabbed with the wrong locale and didn't showed up.

Approach
I added in the document creation another step to always convert the locale to the internal locale, when the path starts with a field and, even though I couldn't find such case, I also added another check that the path should only has the length 3.